Adding WAL encryption (https://redirect.github.com/duckdb/duckdb/pull/17955)
This PR adds WAL encryption. When a database is encrypted, either by

```
ATTACH 'encrypted.db' as encrypted (ENCRYPTION_KEY 'asdf')
```

The WAL files are now encrypted **by default** for encrypted databases.

#### Internals
Every WAL file has a version, and the default version for regular WAL
files is 2. The WAL version for an encrypted WAL file is 3. Instead of
encrypting a WAL file as a whole, we encrypt per entry. Every WAL entry
consist of a length, a checksum and then the actual entry. For entries
encrypted with AES GCM, we keep the length plaintext, then store a
12-byte nonce, encrypt both the checksum and the entry itself and then
store a 16-byte tag. This leads to 28 extra bytes _per entry_ for a WAL
file that is encrypted with GCM.

#### Tests

I've added some test under `test/sql/storage/encryption/wal`. 

#### Debugging

I've added a SETTING: `wal_encryption`. This _could_ be used for
improving performance if one disables WAL encryption. However, this
setting should be carefully used (or not at all). Setting when there
already exist WAL files will lead to corrupted WAL files for encrypted
databases.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0774533b3733f455531c6b56071d18b5e6590802