Unittester: add `on_new_connection` + `on_load` + `skip_tests` options (https://redirect.github.com/duckdb/duckdb/pull/18042)
This PR adds on top of https://github.com/duckdb/duckdb/pull/17972 the
logic needed to actually run end-to-end the test suite while running
custom logic on init / on connection init.
In particular 3 new fields:
* `on_new_connection` for SQL statements to be executed on creating a
new connection (such as `USE db;` that is not persistent cross
connections)
* `on_load` that instruct how to handle sqllogic level `load`
statements. Currently accepts only parameters `` (default) or `skip`,
but could be expanded in the future to allow other modes
* `skip_tests`, a JSON array of tests that have to be skipped. This
allow to encode, within a config, the tests that are expected to fail /
or irrelevant for the configuration.

Those plus the preexisting infrastructure allows to run end-to-end tests
with different configurations, such as the included config with an
encrypted database. Running sqlite or ducklake or else as default
persistent backend is also possible, but left as exercise.

Added configs are checked in at `test/configs/`, and there is a CI job
that execute them.
I think after initial iterations, proper CI setup would be that only
added / modified configs are checked on PRs, while a scheduled job can
then check older configs are still in order.

Contributions by:
* @lnkuiper that did a speed run to change `ParseJSONMap`/`ParseJSON`
that now allow also parsing [nested] arrays, thanks a lot!!
* @Mytherin for feedback
* @ccfelius for testing of initial stage, I took liberty to include a
config with encrypted databases inspired by that

Upstream-commit: https://github.com/duckdb/duckdb/commit/2c35cb7185aa02d5a96a780bc33e8b2f69a56a60