Unittest: Configure skip error messages (https://redirect.github.com/duckdb/duckdb/pull/18087)
One usability feature for config-based usage of unittest is ability to
provide an array that overrides `ignore_error_messages`, meaning that
one can say turn of failures for say "Not Implemented Exception" or
other cases of controlled failure modes. Or maybe if the goal is
stability of the system, most exceptions but for Internal or Fatal ones
can potentially be discarded.

While touching the code, fixed some current limitations:
* summarization would happen before checking whether something is to be
skipped
* in case of mismatched error messages `ignore_error_messages` was not
used
* tests skipped due to `ignore_error_messages` would count as success.
Now they are logged as `skipped`. Of this one I am less sure, and it's
independent, so can be reverted. But my reasoning is that the test is
still not performed to completion, so should probably count as a `skip`

Result summarization will include the cause for the `skip`, like see
last 3 lines:
```
Skipped tests for the following reasons:
require autocomplete: 15
require block_size: 2
require exact_vector_size: 1
require icu: 37
require inet: 1
...
require-env TEST_PERSISTENT_SECRETS_AVAILABLE: 2
skip on error_message matching 'Memory': 396
skip on error_message matching 'could not allocate block of size': 147
skip on error_message matching 'failed to pin block of size': 45
```

I added a test configuration, that is not super relevant, but
demonstrate usage of `skip_error_messages` parameter.

Upstream-commit: https://github.com/duckdb/duckdb/commit/762061c5c0d3b7d92aa494764c980ec9e5e22c0f