[CI] don't run jobs on draft PRs (https://redirect.github.com/duckdb/duckdb/pull/18016)
We'd like to skip running CI on draft PRs and stop running CI on the PRs
that got drafted.

Expected behaviour:

| Event | Current behaviour | Intended behaviour of CI after PR merged |
|---|---|---|
| Open PR | CI starts | CI starts |
| Open **draft** PR | CI starts | CI starts, runs the `check-draft` job
and skips all other jobs |
| Convert to draft | - | CI starts, runs the `check-draft` job and skips
all other jobs; previously triggered CI's got cancelled because of the
same concurrency group |
| Ready for review | CI starts | CI starts; previously triggered CI's
got cancelled because of the same concurrency group |
| Push to open PR | `DraftPR.yml` turns it into a draft; previous CI
keeps running | CI starts; `DraftPR.yml` turns it into a draft; then it
becomes the same as `Convert to draft` |
| Push to **draft** PR | - | CI starts, runs the `check-draft` job and
skips all other jobs |

Expected effect of the PR is stop running CI on branches that are
outdated or not ready to be evaluated yet.

Upstream-commit: https://github.com/duckdb/duckdb/commit/e4d1d60ae688ade04fe97ac3d711e2ccb0c0fcc2