Add support for `MERGE INTO` (https://redirect.github.com/duckdb/duckdb/pull/18135)
This PR adds support for the SQL standard `MERGE INTO` statement.
The `MERGE INTO` statement allows the user to perform inserts, updates
or deletes conditionally depending on what data is present in the source
or target.
It is useful to e.g. allow for upserts without requiring an index to be
present.

### Examples

Suppose we keep track of a list of items in a warehouse, and we buy or
sell items.


```sql
CREATE TABLE item(item_name varchar, count int);
INSERT INTO item VALUES ('book', 3), ('computer', 5);
```

When buying more items, we either insert them to the table (if they
don't exist yet), or we increment the count.
We can model this with a `MERGE INTO` statement with the following
clauses:

```sql
WITH buy(item_name, count) AS (
	VALUES ('book', 2), ('phone', 3)
)
MERGE INTO item USING buy ON (item.item_name = buy.item_name)
WHEN MATCHED THEN UPDATE SET count = item.count + buy.count
WHEN NOT MATCHED THEN INSERT;

FROM item;
┌───────────┬───────┐
│ item_name │ count │
│  varchar  │ int32 │
├───────────┼───────┤
│ book      │     5 │
│ computer  │     5 │
│ phone     │     3 │
└───────────┴───────┘
```

When selling items, we either delete the item if there are none left, or
we decrement the count.
In addition, we want to print an error message if we are selling items
we don't have.

```sql
WITH sell(item_name, count) AS (
	VALUES ('book', 5), ('computer', 2)
)
MERGE INTO item USING sell ON (item.item_name = sell.item_name)
WHEN MATCHED AND sell.count > item.count THEN ERROR CONCAT('Trying to sell ', sell.count, ' copies of item ', item.item_name, ' but we only have ', item.count)
WHEN MATCHED AND sell.count = item.count THEN DELETE
WHEN MATCHED THEN UPDATE SET count = item.count - sell.count
WHEN NOT MATCHED THEN ERROR CONCAT('Trying to sell item ', sell.item_name, ', but we don''t have it in stock');

FROM item;
┌───────────┬───────┐
│ item_name │ count │
│  varchar  │ int32 │
├───────────┼───────┤
│ computer  │     3 │
│ phone     │     3 │
└───────────┴───────┘
```

If we try to sell the same items again, we get the following error as we
ran out of books to sell:

```
Merge error condition WHEN NOT MATCHED: Trying to sell item book, but we don't have it in stock
```

#### Syntax

The complete syntax for `MERGE INTO` is as follows:

```sql
[ WITH WithQuery ]
MERGE INTO TargetTableName [ [ AS ] Alias ]
USING TableRef [ [ ON JoinCondition ] / [ USING UsingColumns ] ]
MatchClauses
```

###### MatchClauses

```sql
- WHEN MATCHED [ AND condition ] THEN MergeAction
- WHEN NOT MATCHED BY SOURCE [ AND condition ] THEN MergeAction
- WHEN NOT MATCHED [ BY TARGET ] [ AND condition ] THEN MergeAction
```

###### MergeAction

```sql
- INSERT [ ( ColumnList ) ] VALUES ( ValueList )
- DELETE
- UPDATE SET SetList

- UPDATE [ BY NAME / BY POSITION ] [ SET * ]
- INSERT [ BY NAME / BY POSITION] [ * ]
- DO NOTHING
- ERROR [ error_message ]
```


We support all SQL standard aspects of the `MERGE INTO` operator,
together with several extensions also supported by other systems, in
addition to some of our own. The extensions we support are:

* `WHEN NOT MATCHED BY SOURCE / BY TARGET` - also supported by
[Postgres](https://www.postgresql.org/docs/current/sql-merge.html),
[DataBricks](https://docs.databricks.com/aws/en/sql/language-manual/delta-merge-into)
* `DO NOTHING` as an explicit action - also supported by
[Postgres](https://www.postgresql.org/docs/current/sql-merge.html)
* `UPDATE SET * / INSERT *` - also supported by
[DataBricks](https://docs.databricks.com/aws/en/sql/language-manual/delta-merge-into)
* `UPDATE / INSERT BY NAME` - DuckDB extension
* `ERROR [ error_message ]` - DuckDB extension
* `USING` clause for join - DuckDB extension



#### WHEN NOT MATCHED BY SOURCE / WHEN NOT MATCHED BY TARGET

By default, and as specified in the SQL standard, `WHEN NOT MATCHED` is
triggered for a row in the source that is not found in the target table
we are merging into. However, in certain cases, we also want to trigger
an action when there is a row in the target that has not found a match
in the source. This can be specified using the `WHEN NOT MATCHED BY
SOURCE` condition.

For example, if we want to use the `MERGE INTO` statement to make two
tables fully equivalent, we can use the following merge into statement:

```sql
CREATE TABLE target_tbl(id int, count int);
CREATE TABLE source_tbl(id int, count int);
INSERT INTO target_tbl VALUES (1, 10), (3, 7);
INSERT INTO source_tbl VALUES (1, 5), (2, 12);

MERGE INTO target_tbl USING source_tbl USING (id)
WHEN MATCHED THEN UPDATE
WHEN NOT MATCHED BY SOURCE THEN DELETE
WHEN NOT MATCHED BY TARGET THEN INSERT;

FROM target_tbl;
┌───────┬───────┐
│  id   │ count │
│ int32 │ int32 │
├───────┼───────┤
│     1 │     5 │
│     2 │    12 │
└───────┴───────┘
```

This statement effectively means:

* If we find a row with the same id in both `source` and `target`,
update the values in `target` so that they are equivalent to that in
`source`
* If a row exists in `source` that does not exist in `target`, insert it
* If a row exists in `target` that does not exist in `source`, delete it

The presence of the various `WHEN NOT MATCHED` clause effectively
determines the join type we need to use to perform the merge:

* `WHEN NOT MATCHED BY TARGET`: LEFT OUTER JOIN
* `WHEN NOT MATCHED BY SOURCE`: RIGHT OUTER JOIN
* Both: FULL OUTER JOIN
* Neither: INNER JOIN

#### INSERT [ BY NAME / BY POSITION ], UPDATE [ BY NAME / BY POSITION]


`INSERT` and `UPDATE` by default need a list of clauses, e.g. `INSERT
VALUES (col1, col2)` or `UPDATE SET col1 = val, col2 = val`.

As an extension, DuckDB also supports automatically generating these
from the source table.
This is similar to how `INSERT` from a query works, e.g.:

```sql
MERGE INTO target_tbl USING source_tbl ON (false)
WHEN NOT MATCHED THEN INSERT;

-- is equivalent to
INSERT INTO target_tbl FROM source_tbl;
```

```sql
MERGE INTO target_tbl USING source_tbl ON (false)
WHEN NOT MATCHED THEN INSERT BY NAME;

-- is equivalent to
INSERT INTO target_tbl BY NAME FROM source_tbl;
```

This is also supported for `UPDATE`, where we can do `UPDATE` or `UPDATE
BY NAME` to update all columns in the target table either by position or
by name from the source table.

Upstream-commit: https://github.com/duckdb/duckdb/commit/acf9008d47a132df5eefc6621ecf6bbd7b9a0790