Detect when updates have no effect, and skip performing the actual updates if we encounter these nop updates (https://redirect.github.com/duckdb/duckdb/pull/18144)
Currently when running a "nop" update, like:

```sql
UPDATE tbl SET col=col;
```

We perform the actual update. This is wasteful. While the above example
is clearly rather pointless - nop updates can be performed accidentally
quite easily, especially when updating many columns.

This PR adds in detection for when an update leaves a value unchanged -
and when it leaves a value unchanged, it will skip the update for that
value. In the best case scenario that means we can skip applying the
entire update (and also skip subsequent operations, such as rewriting
row groups).

Upstream-commit: https://github.com/duckdb/duckdb/commit/5f23f06bf70358fbf86ff32a7665a0370ad0c607