fix bug with allowed_paths (https://redirect.github.com/duckdb/duckdb/pull/18176)
### Problem
The `allowed_directories` values was not searched correctly leading to
permissions errors in certain cases.

### Fix
Current behaviour tried to use an `std::set ` to efficiently do prefix
matches. The implementation was actually wrong though. It used the
`set::upper_bound` and `set::lower_bound` to limit which paths should be
compared for a prefix match. However this does not work. These functions
return the exact same value unless the exact path you search for is in
the set.

I think the `set` can not be used for efficient prefix matching like
this. Because this is not performance critical code anyway I've just
switched to a O(N) search over the set of prefixes. Proper solution
would be to use a trie, but we don't have a simple implementation handy
afaik.

Made a small tweak to one of the tests to trigger this error and confirm
the fix

Upstream-commit: https://github.com/duckdb/duckdb/commit/97d22c2a22a58fcbb7878c164ea74ba6568f5376