Reduce lock contention for the instance cache (https://redirect.github.com/duckdb/duckdb/pull/18079)
Previously, the instance cache lock was held for the entire duration of
a new DuckDB instance's creation. This blocked the creation of any other
instances, causing significant contention, especially for slower
connections like those to MotherDuck.

This PR resolves the issue by introducing futures to the cache. Now, a
future is immediately placed in the cache and the lock is released. The
database instance is then created after releasing the lock, with the
result being used to populate the future upon completion. This allows
multiple instances to be initialized concurrently without blocking.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c677fa112c5597c16cd83ed24739129696d38d53