Improve Parquet reader `NULL` statistics and compress all-`NULL` columns using `CompressedMaterialization` (https://redirect.github.com/duckdb/duckdb/pull/18217)
PR summary:
1. We already derive `BaseStats::CanHaveNull()` for Parquet columns
(whether there are `NULL` values), but not yet
`BaseStats::CanHaveNoNull()` (whether there are any non-`NULL` values).
This PR implements the latter for Parquet column statistics.
2. When reading multiple Parquet files, we bail on combining column
statistics, as this would require parsing all metadata while binding the
query. However, when `union_by_name=true`, we parse all metadata, but we
still bail on combining column statistics. This PR implements combining
column statistics from multiple Parquet files when `union_by_name=true`.
3. The `CompressedMaterialization` optimizer creates projections to
reduce the width of columns when materializing them in blocking
operators like join, aggregation, and sort. We do this for integral
values based on the min/max statistics, and for string values based on
the maximum string length statistic. This PR adds support for
compressing columns to a `UTINYINT` when we can prove that all values in
the column are `NULL`, which can drastically reduce the size of sparse
data.

It is not often the case that all values in a column are `NULL`, but
when working with a subset of sparse data, which is sometimes the case
when working with a few (out of many) Parquet files, this optimization
can sometimes drastically reduce the memory usage. I've added a somewhat
ridiculous test that builds a join hashtable on 400 all-`NULL` `VARCHAR`
columns, which would've used ~8 GB of RAM, but now compresses down to
~0.5 GB.

Upstream-commit: https://github.com/duckdb/duckdb/commit/a6d1fbf4088bbe68eaa1dc66082f9b4f66cdad19