[Parquet] Add read support for the `VARIANT` LogicalType (https://redirect.github.com/duckdb/duckdb/pull/18187)
This PR adds support for the (unshredded) VARIANT type in Parquet, as
defined by
https://github.com/apache/parquet-format/blob/master/VariantEncoding.md

We transform this binary encoding to JSON, this does not require the
JSON extension, but the JSON extension is required to interact with the
result.

Tests have been generated using [this
test](https://github.com/apache/iceberg/blob/main/parquet/src/test/java/org/apache/iceberg/parquet/TestVariantWriters.java)
from the apache/iceberg repo, which is one of the first systems to
implement support for this new type.

We introduce a new setting: `variant_legacy_encoding` to enable the
parquet reader to identify columns with the shape: `STRUCT("metadata"
BLOB, "value" BLOB)` as a Variant.
The spec says that the parquet "group" should be annotated with the
`VariantType` logical type but Iceberg's implementation doesn't do this
yet, so we're forced to identify it structurally for the time being.

Upstream-commit: https://github.com/duckdb/duckdb/commit/59b01088c9ccb4c451200f6303b75be09b714530