Better `NULL` handling in `TupleDataLayout` (https://redirect.github.com/duckdb/duckdb/pull/18069)
DuckDB already handles `NULL`s efficiently (1 check per vector, usually,
if there are no `NULL`s), which is a lot less brittle than handling
`NULL`s once for the entire query plan.

However, when we materialize data in row oriented layout, we go back to
checking `NULL`s once per value. This is wasteful, especially if we know
from the statistics that there are no `NULL`s in the query plan.

This PR implements `NULL` handling for the `TupleDataLayout`, and
propagates the `NULL` statistics information to the aggregation operator
in the optimizer. We can add this for joins too in a future PR.

I've measured up to 10% performance improvement for simple aggregations,
but it all depends on the query, of course. If there are `NULL`s, the
performance is about the same, and if there long strings, for example,
performance is dominated by handling those.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1b6c85a228cb2c12a4bf7280c8f10b0227df30d7