Add support for geoarrow encoded geometries in geoparquet files. (https://redirect.github.com/duckdb/duckdb/pull/17942)
This PR starts working on adding supporting geoarrow geometry types (see
discussion at https://github.com/duckdb/duckdb/discussions/17618).

This PR enables geoarrow columns to be read as structures. However, it
does not mark the columns as Geometry columns.

Could I get some guidance on next steps? Should the returned column have
a Logical type of Geometry and a physical type of Struct. Or is that not
required?

Upstream-commit: https://github.com/duckdb/duckdb/commit/e7af00a34231f67369eff980e69aa8d60a39e6ea