Add type safety to `FlatVector::GetData<T>`, `ConstantVector::GetData<T>` and `UnifiedVectorFormat::GetData<T>` (https://redirect.github.com/duckdb/duckdb/pull/18256)
This PR adds type safety to the various `GetData<T>` methods on our
`Vector` type. We use `GetTypeId<T>()` to obtain the `PhysicalType` that
corresponds to the type we are trying to access based, and then compare
it to the physical type that is actually stored in the vector. If the
types don't match, we throw an internal exception. So if we e.g. use
`FlatVector::GetData<string_t>(vec)` on a vector of type `INTEGER` - an
exception is thrown.

Each type has a single physical type it corresponds to - i.e. `int64_t`
corresponds to `PhysicalType::INT64`. One exception is made for
`PhysicalType::BOOLEAN` - we allow this type to be accessed also using
`int8_t` and `uint8_t` which happens in many places in the code.

In a number of places we "reinterpret" data stored in a vector - i.e. we
intentionally access `UINT64` for a `DOUBLE` vector, or `HUGEINT` for a
`LIST` vector. We have now changed these places to use `GetDataUnsafe`.
We could also consider in the future adding a `ReinterpretData` method
that instead checks that the byte sizes of types are compatible (i.e.
that would allow a `DOUBLE` to be read as `UINT64`, but not as `UINT32`)
- but we leave that for a future iteration.


#### Extension Types
While there is a big type switch in `GetTypeId<T>`, extensions can also
make use of their own types and plug it into the infrastructure using
the following overload as used in e.g. spatial:

```cpp
template <>
inline PhysicalType GetTypeId<geometry_t>() {
	return PhysicalType::VARCHAR;
}
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/906d8481267f9a9a2102f2faeb72092ad30cadbe