[Parquet] Add read support for the `VARIANT` LogicalType (with shredded encoding) (https://redirect.github.com/duckdb/duckdb/pull/18224)
This PR builds on top of https://redirect.github.com/duckdb/duckdb/pull/18187 which added support for the "unshredded"
version
(https://github.com/apache/parquet-format/blob/master/VariantEncoding.md)

We add support for the "shredded" version
(https://github.com/apache/parquet-format/blob/master/VariantShredding.md),
which is an optimization at the storage level to store the most common
representation of the column (and fields of the column, if it's an
object) in a columnar format with a strong type, outside of the
"unshredded" binary-encoded representation. (see the parquet spec for
more details)

This contains all of the generated parquet files from apache/iceberg's
[TestVariantWriters](https://github.com/apache/iceberg/blob/main/parquet/src/test/java/org/apache/iceberg/parquet/TestVariantWriters.java)
test file.
These are rewritten with to add the `logicalType` VARIANT to the
top-level struct of the Variant columns.

The variants containing Timestamp Nanos TZ columns still require the
`variant_legacy_encoding` setting, which will stay true until the
variants written by apache/iceberg contain the `VariantType` logical
type.

### Implementation Details

For future reference:
Shredded variants are recursive, and by that I mean that the
`STRUCT(value BLOB, typed_value <type>)` can appear at all levels until
the `typed_value` is no longer a `LIST` or `STRUCT`, or `typed_value` is
omitted entirely.

`value` is always present.
`typed_value` is optional.

`metadata` is only present at the top level, but its value is relevant
for the entire row. (so we forward it to recursive `Convert` calls
through a ConstantVector)

`typed_value` is checked first, if it's not null.
When `typed_value` is null, `value` has to not be null[1].
When `typed_value` is not null, `value` has to be null[2].

[1]: This is not true for fields of an Object, if both are null it
signals that the field is not present in the Object.
[2]: This is not true for Objects, even if `typed_value` is not null,
`value` can be null, to contain a binary Object with the remaining
fields that aren't shredded on.

### Future Work

- Native DuckDB Variant type, so we don't have to output JSON.
- Projection pushdown in the form of JSON Paths (our existing projection
pushdown works with ColumnIndex, which can't apply, because we can't
assign an id to arbitrary fields).
- Emitting stats when possible (if the Variant is fully shredded)

Upstream-commit: https://github.com/duckdb/duckdb/commit/d7f955fd3cba2f4d926274d45e0b12ac9b615d98