Expanded autocomplete suggestions (https://redirect.github.com/duckdb/duckdb/pull/18243)
This PR introduces autocomplete suggestions for: 
- Pragma
- Settings (Using `SET` syntax)
- Scalar functions
- Table functions

This resolves some open TODOs. 

The PR doesn't introduce any changes to the grammar and is separate from
https://github.com/duckdb/duckdb/pull/18221

To provide these suggestions, I needed a way to get a complete list of
each type of entry. My current approach adds a new method to the
`Catalog` class (`GetAllEntries`). This works, but I don't want to
convolute the class with unneeded methods, so happy to hear feedback
regarding this :)

Upstream-commit: https://github.com/duckdb/duckdb/commit/6d74aa8e5f51d53c8c656d60e17d5c4f6c758663