Use `FromEpochSeconds` instead of `FromTimeT` in `FileSystem::GetLastModifiedTime` (https://redirect.github.com/duckdb/duckdb/pull/18281)
This should fix the initial issue in
https://github.com/duckdb/duckdb/issues/18212, but not the
`HUGEINT`-related issue (in a comment, not the initially created issue).

I could reproduce the wrong timestamp behavior locally, but it was
different than the behavior in the issue because I'm in UTC PLUS 2,
while the issue creator is in UTC MINUS something.

Upstream-commit: https://github.com/duckdb/duckdb/commit/0d0c0190c4d31a40d27387c45c0b24440c068e52