Add target for installing Python deps. (https://redirect.github.com/duckdb/duckdb/pull/18285)
Add a target to make installing Python deps more systematic. Maybe it
makes sense to have `generate-files` error out the Makefile command if
any part of it fails as well, because right now the error gets buried in
logs.

`PYTHON` as an envvar in the `Makefile` allows overriding the use of
`python3` if using `pyenv` or similar which uses `python` instead.

```
PYTHON=python make generate-files
```

Some of this could be simplified with `uv` but there's so few deps it's
probably not worth the overhead.

Upstream-commit: https://github.com/duckdb/duckdb/commit/199e5a60de7f595bc1393e31fad9707cf45db3bc