add the from-table-function as parameter to copy-from-bind (https://redirect.github.com/duckdb/duckdb/pull/18004)
MotherDuck overloads TableFunctions (and CopyFunctions) but would look
to to so in a more inobtrusive manner, changing nothing if the functions
execute locally. This change allows the copy-from-bind to make changes
to the from-table-function in case it executes remotely (more precisely,
it will change some callback hooks to RPC stubs).

Upstream-commit: https://github.com/duckdb/duckdb/commit/61f07c3221e01674d8fe40b4a25364ba2f3159a7