Extend PEG parser grammar (https://redirect.github.com/duckdb/duckdb/pull/18221)
This PR extends the grammar used by the autocomplete functions to work
for (almost) all SQL queries in the `test` folder and out-of-tree `test`
folders. I realise it has become quite a large PR, but does make the PEG
parser more complete, moving forward to an extensible parser.

It adds support for:  
- Unicode spaces: These were initially not filtered out and caused the
PEG parser to crash. They are now replaced by normal spaces if they are
detected. This is similar in behaviour to the PostgreSQL parser (`const
string &sql_ref = StripUnicodeSpaces(sql, clean_sql) ? clean_sql :
sql;`)
- Dollar-quoted strings. 
- New syntax
- Better keyword handling (more inline with PostgreSQL parser)
- Tests for autocomplete on `ALTER` columns. 

## New syntax 
- `MERGE INTO`
- `LAMBDA`
- `ATTACH`
- `CREATE INDEX`
- Loading and installing extensions
- More elaborate type support, such as `IntervalType`
- Vacuum

## Keywords
The old PEG parser still relied on the PostgreSQL keyword `.list` files
to determine the category of the keyword. For the PEG parser, I copied
these `.list` files and placed them in the `keywords` folder. Based on
the file they are in, they are categorized in a `keyword_map` that is
used during parsing. This removes the dependency on the PostgreSQL
parser, but it adds the complexity that there are now two of these sets
of keywords (for PEG parser and PostgreSQL parser).

Normally table and schema names are not allowed to be reserved keywords.
Except if the catalog name is specified before the schema name, or the
schema name before the table name. For instance, `AND` is a reserved
keyword:
`USE catalog_name.and` ✅ 
`USE and` ❌ 

`FROM schema_name.and` ✅ 
`FROM and` ❌ 

To handle the difference between these cases, and still keep the table
and schema name suggestions, I added `RESERVED_TABLE_NAME` and
`RESERVED_SCHEMA_NAME` to the `SuggestionState`. In the
`IdentifierMatcher`, these can then match on a reserved keywords,
whereas `TABLE_NAME` and `SCHEMA_NAME` cannot. The grammar rule related
to this is:
`QualifiedName <- (CatalogQualification ReservedSchemaQualification
ReservedTableName) / (SchemaQualification ReservedTableName) /
(Identifier / StringLiteral)`

When providing suggestions, we then still have the context that we want
to provide a suggestion for either a table name or schema name.

The `keyword_map` is automatically generated using the
`extension/autocomplete/inline_grammar.py` script, which generates the
entire grammar. I also added a script: `scripts/build_peg_grammar.sh`
which generates both the `inlined_grammar.hpp` and
`inlined_grammar.gram` files.

### Changes to keywords
I moved some of the keywords, such as `MAP` out of the
`func_name_keywords.list` file. `MAP` is categorized as both a
`column_name` as well as `func_name` in the PostgreSQL parser. In the
grammar file it is categorized as: `PG_KEYWORD("map", MAP,
TYPE_FUNC_NAME_KEYWORD)`.

My confusion lies in the following query: `CREATE TABLE MAP(i BIGINT);`,
which is a valid query. However, a table name may only be (1) an
Identifier (no category of keyword) (2) an unreserved keyword (3) a
column_name keyword. Seeing as map is categorized as
`TYPE_FUNC_NAME_KEYWORD`, I don't understand how this query passes with
the PostgreSQL parser, but it does. To solve this in the PEG parser, I
removed `MAP` from the `func_name_keywords.list` file, but I don't think
this is the best solution. Happy to hear feedback regarding this.


## Failing test cases: 
The following test cases are excluded from the new PEG parser:
1. `PREPARE s1 AS SELECT ?1`
Reason for Exclusion: This query relies on a parsing behavior where a
number following a ? parameter is treated as a column alias.
Explanation: The standard parser tokenizes ?1 as two separate tokens: a
positional parameter (?) and an integer literal (1). The literal is then
interpreted as an alias, making the query equivalent to SELECT ? AS "1".
This behavior is specific to the interaction between the tokenizer and
the standard SQL parser. But to me this seems more like a bug. Normally
only `$1` or `?` should accepted.
 
2. `SET s3_use_ssl=${DUCKDB_S3_USE_SSL}`
This seems like a weird interaction where we are missing `''` around the
`${}` and somehow the SQLLogicTest framework can handle this, but trying
this in the CLI seems to cause a syntax error.
    
## Script changes
I also extended the `scripts/test_peg_parser.py` script that tests the
PEG parser in two ways:
1. By making it multithreaded (because I was tired of waiting around for
it to finish) when setting the `--no-exit` flag.
2. `--include-extensions` Downloads the test files for all the
out-of-tree extensions and stores them under `extension-test-files` in
the project root directory (to avoid downloading these files every
time).

Upstream-commit: https://github.com/duckdb/duckdb/commit/566e5fd29b68268198007fe1518d979803787b71