Re-use table metadata when table is not altered during checkpoint (https://redirect.github.com/duckdb/duckdb/pull/18390)
Currently during a checkpoint we rewrite all metadata associated with
all tables. For larger tables, the metadata can be relatively large.
This causes checkpoints on larger databases to always take a base amount
of time - regardless of what has actually changed.

The WAL mitigates this to some extend by buffering changes - but
rewriting metadata is still wasteful:
* It causes checkpoints to take an increasing amount of time based on
database size, regardless of change size
* It causes checkpoints to write a lot of data to the database, even if
little changed

This PR partially mitigates this issue by re-using table metadata for
tables that are fully unchanged. During checkpointing, we check if any
changes have been made to a table. If no changes have been made, the
entire checkpoint for that table is short-circuited and we write a
pointer to the previous metadata.

Note that this is not entirely free. We still need to read parts of the
previous metadata in order to figure out which blocks are used by the
previous metadata. This is so that we can correctly label these blocks
as still in-use, so that they do not get accidentally garbage collected.

### Performance

Consider a database storing TPC-H SF100. The biggest table (lineitem)
has 600M rows.

```sql
CALL dbgen(sf=100);
```

Now insert a single row into a tiny table (`nation`) and checkpoint. We
measure two times: the time of the `CHECKPOINT` command, and the full
runtime of opening the database, running the insert + checkpoint, and
closing the database.

```sql
INSERT INTO nation VALUES (26, 'New Country', 0, '');
CHECKPOINT;
```

| Operation  | v1.3.2 |  New  |
|------------|--------|-------|
| Checkpoint | 0.53s  | 0.07s |
| Full       | 0.64s  | 0.1s  |


Note that there is still a reasonably-sized cost here. The main
remaining cost is that we are not serializing the exact blocks the row
groups refer to, meaning we need to do a partial deserialization to
figure this out. If we were to serialize the exact blocks of each row
group we would be able to reduce the checkpoint time by another ~2.5X
based on rudimentary experiments

Upstream-commit: https://github.com/duckdb/duckdb/commit/df7f4fa0e56f81950e8e84a314a2a00e0acf3ff5