CI: Fix Discussion mirroring (https://redirect.github.com/duckdb/duckdb/pull/18397)
This PR fixes discussion mirroring.

The `gh` CLI tool has limited support to work with discussions
(https://github.com/cli/cli/discussions/4212), so the recommendation is
to script using GraphQL. However, we actually do not need to do much
with the discussions to mirror them: we simply need to open an internal
issue. There is no need to adjust the labels of discussions.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c4dc9ee5be0ec6498d41ac8d8aa757a9cd6d9660