Store extra metadata blocks in RowGroupPointer, and only flush dirty Metadata blocks (https://redirect.github.com/duckdb/duckdb/pull/18398)
Follow-up from https://github.com/duckdb/duckdb/pull/18395

This PR adds storage for `extra_metadata_blocks` in the row-group
metadata when the latest storage version (v1.4.0) is used. This is a
list of metadata blocks that are referenced by the row group metadata,
but **not** present in the list of data pointers (`data_pointers`).
These blocks can be present when either (1) columns are very wide due to
e.g. being deeply nested, or (2) the final column pointer "crosses" the
metadata block threshold, as `data_pointers` points only to the
beginning of the column metadata.

Usually this list is empty - and therefore storing this does not take up
a lot of extra storage space. The presence of this list allows us to
more efficiently re-use metadata as we know exactly which metadata
blocks a row group points to, without having to do any additional
deserialization.

### Only Flush Dirty Metadata blocks

Previously our metadata manager would flush all metadata blocks,
incurring a lot of unnecessary I/O now that we are doing a lot of
metadata re-use. This PR reworks this so that we keep track of which
blocks are dirty and only flush the dirty blocks.

### Performance

Running the benchmark in https://github.com/duckdb/duckdb/pull/18395
again, we now get the following timings:

| Operation  | v1.3.2 |  Re-Use  |  New  |
|------------|--------|-------|-------|
| Checkpoint | 0.5s   | 0.11s | 0.04s |
| Full       | 0.63s  | 0.13s | 0.07s |

Upstream-commit: https://github.com/duckdb/duckdb/commit/0769417a29fc710aa2d9535fc281c1723def3f6e